Began a few restructurings of the codebase
