+++ OpModes/motionLibrary/Main.java
@@ -1,0 +1,9 @@
+package org.firstinspires.ftc.teamcode.OpModes.motionLibrary;
+
+import java.util.ArrayList;
+import org.firstinspires.ftc.teamcode.Other.Backend.MadHardware;
+import com.qualcomm.robotcore.util.ElapsedTime;
+
+public class Main {
+  
+}
